requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==0.3.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
